refactor: Dealer 추가
