+++ src/main/kotlin/blackjack/dto/PlayerDto.kt
@@ -1,7 +1,7 @@
 package blackjack.dto
 
-import blackjack.domain.player.Player
+import blackjack.domain.player.Gamer
 
 data class PlayerDto(val name: String, val cards: List<String>) {
-    constructor(player: Player) : this(player.name, player.cards.toView())
+    constructor(player: Gamer) : this(player.name, player.cards.toView())
 }
