+++ src/main/kotlin/blackjack/dto/ResultDto.kt
@@ -1,7 +1,7 @@
 package blackjack.dto
 
-import blackjack.domain.player.Player
+import blackjack.domain.player.Gamer
 
 data class ResultDto(val name: String, val cards: List<String>, val score: Int) {
-    constructor(player: Player) : this(player.name, player.cards.toView(), player.cards.score.value)
+    constructor(player: Gamer) : this(player.name, player.cards.toView(), player.cards.score.value)
 }
