+++ src/test/kotlin/blackjack/domain/player/PlayerTest.kt
@@ -19,15 +19,7 @@ internal class PlayerTest {
     fun `이름이 같을 경우 같은 플레이어로 취급한다`() {
         assertThat(Player("자손", cards)).isEqualTo(Player("자손", cards))
     }
-
-    @ParameterizedTest
-    @MethodSource("provideDealerCards")
-    fun `딜러 카드의 점수기 16점 보다 높은 경우 카드를 더이상 가질 수 없다`(cards: Cards, expected: Boolean) {
-        val player = Player("자손", cards)
-        val result = player.isTakeableDealer()
-        assertThat(result).isEqualTo(expected)
-    }
-
+    
     @ParameterizedTest
     @MethodSource("provideCards")
     fun `플레이어 카드의 점수가 블랙잭(21점)보다 높거나 같은 경우 카드를 더이상 가질 수 없다`(cards: Cards, expected: Boolean) {
@@ -62,26 +54,6 @@ internal class PlayerTest {
                         Card(Suit.DIAMOND, Denomination.JACK)
                     ),
                     true
-                )
-            )
-        }
-
-        @JvmStatic
-        private fun provideDealerCards(): Stream<Arguments> {
-            return Stream.of(
-                Arguments.of(
-                    Cards(
-                        Card(Suit.HEART, Denomination.JACK),
-                        Card(Suit.DIAMOND, Denomination.SIX)
-                    ),
-                    true
-                ),
-                Arguments.of(
-                    Cards(
-                        Card(Suit.HEART, Denomination.JACK),
-                        Card(Suit.HEART, Denomination.SEVEN)
-                    ),
-                    false
                 )
             )
         }
